Cloud_2/requirements.txt
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
coards==1.0.5
Extract==1.0
google-api-core>=2.14.0
google-auth[requests]>=2.27.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
idna==3.10
lxml==6.0.0
numpy==2.3.1
oauthlib==3.3.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydap==3.5.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
rarfile==4.2
requests==2.32.4
requests-cache==1.2.1
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.16.0
six==1.17.0
soupsieve==2.7
typing-inspection==0.4.1
typing_extensions==4.14.1
url-normalize==2.2.1
urllib3==2.5.0
WebOb==1.8.9
google-cloud-pubsub>=2.0.0
sqlanydb~=1.0.14